# Measure with best available precision