Fixes #207

* Introduce a separate measurer to use during warmup which uses the native time
* add nano seconds to our list of available units
* change up display of times in places