@PragTob 2018-05-10
ah jeez coveralls is failing now because in the runner all the repetition code is now untested as only the linux builds report to it :'(

We could adjust coveralls max coverage drop or extract the code and see what's up.