@PragTob 2018-05-11
Rebuilding appveyor, it seems to have problems to access hex.pm