@PragTob 2018-05-12
Ok I wrote the beginning of a little benchmark I wanted to do for board games and compared this against master (yes I know lists are bad for this, end goal is to compare vs. maps, tuples, records):

```elixir
defmodule BoardBuilder do
  def list(size) do
    range = 1..size
    Enum.map(range, fn _ ->
      Enum.map(range, fn _ -> :value end)
    end)
  end
end

alias Benchee.Utility.RepeatN

list = BoardBuilder.list(19)

Benchee.run(
  %{
    "get(0, 0)" => fn -> list |> Enum.at(0) |> Enum.at(0) end,
    "get(max, max)" => fn -> list |> Enum.at(18) |> Enum.at(18) end
  }
)
```

this branch: 

```
Name                    ips        average  deviation         median         99th %
get(0, 0)            5.86 M      170.62 ns ±11344.34%         150 ns         226 ns
get(max, max)        2.08 M      481.84 ns  ±2347.39%         459 ns         693 ns

Comparison: 
get(0, 0)            5.86 M
get(max, max)        2.08 M - 2.82x slower
```

master:

```
# fast warning with I think even 100 repetition

Name                    ips        average  deviation         median         99th %
get(0, 0)           12.25 M      0.0816 μs   ±292.68%      0.0800 μs       0.130 μs
get(max, max)        2.62 M        0.38 μs     ±9.06%        0.38 μs        0.44 μs

Comparison: 
get(0, 0)           12.25 M
get(max, max)        2.62 M - 4.68x slower
```

so I decided to put in some manual repetition:

```elixir
defmodule BoardBuilder do
  def list(size) do
    range = 1..size
    Enum.map(range, fn _ ->
      Enum.map(range, fn _ -> :value end)
    end)
  end
end

alias Benchee.Utility.RepeatN

list = BoardBuilder.list(19)

Benchee.run(
  %{
    "get(0, 0)" => fn -> RepeatN.repeat_n(fn -> list |> Enum.at(0) |> Enum.at(0) end, 100) end,
    "get(max, max)" => fn -> RepeatN.repeat_n(fn -> list |> Enum.at(18) |> Enum.at(18) end, 100) end
  }
)
```

Then my results are about the same as on master but more precise:

```
Name                    ips        average  deviation         median         99th %
get(0, 0)          119.71 K        8.35 μs    ±89.99%        8.25 μs       10.35 μs
get(max, max)       26.03 K       38.42 μs     ±8.97%       38.23 μs       40.53 μs

Comparison: 
get(0, 0)          119.71 K
get(max, max)       26.03 K - 4.60x slower
```

It makes it seem to me like the **overhead from the function call from the function call is bigger than the one from the repetition** - i.e. I feel that the repeated samples are somewhat more realistic :thinking: 

Now let me set an empty function call as the baseline we get these results:

```
Name                    ips        average  deviation         median         99th %
fn -> nil end       12.62 M       79.22 ns    ±34.72%          78 ns         103 ns
get(0, 0)            5.48 M      182.34 ns ±10699.05%         163 ns         263 ns
get(max, max)        2.04 M      489.85 ns  ±2556.78%         461 ns         698 ns

Comparison: 
fn -> nil end       12.62 M
get(0, 0)            5.48 M - 2.30x slower
get(max, max)        2.04 M - 6.18x slower
```

Let me deduct the empty  average then we end up with: 103 and 410 which says that get(max, max) is about 4 times slower than my get(0, 0). If I do the same using the median (flipping the bird to outliers) We get 85 and 383 respectively arriving at ~ 4.5 slower

So now the question is - **which one of these values is most trust worthy?**
* 2.8 (original nanosecond branch)
* 4.0 (nanosecond branch trying to account for function call overhead with average)
* ~4.5-4.7 (master and manual repetition on nanosecond, nanosecond branch accounting for overhead using median)